Fixed changes according to PR
