+++ app/forms.py
@@ -9,7 +9,6 @@ from app.models import Users
 class FindTeam(FlaskForm):
     year = SelectField('Year',choices=[], validators=[DataRequired()],id="year-dropdown")
     team = SelectField('Team Name', choices=[],validators=[DataRequired()],validate_choice=False,id="team-dropdown")
-
     submit = SubmitField('Find Team Roster')
 
 class ImmaculateGridInput(FlaskForm):
